+++ src/CMakeLists.txt
@@ -366,7 +366,7 @@ list(APPEND SOURCE_FILES
         displayapp/DisplayApp.cpp
         displayapp/screens/Screen.cpp
         displayapp/screens/Tile.cpp
-        displayapp/screens/ImageView.cpp
+        displayapp/screens/FileView.cpp
         displayapp/screens/Gallery.cpp
         displayapp/screens/InfiniPaint.cpp
         displayapp/screens/Paddle.cpp
