+++ src/displayapp/screens/Gallery.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "displayapp/screens/Screen.h"
-#include "displayapp/screens/ImageView.h"
+#include "displayapp/screens/FileView.h"
 #include "displayapp/DisplayApp.h"
 #include <lvgl/lvgl.h>
